Warn about ambiguous include paths
Each include path should be unique in the sense that it is
not a subpath of another include. Otherwise, the same headers can
be included with different paths, which may confuse users
or create build failures if the exported include path differs
from the one that is used internally.

catkin_lint detects these ambiguities now.

Fixes #5