Add #catkin_lint pragmas to parser
Currently, there is one pragma implemented:
#catkin_lint: ignore ID [ID [...]]

This pragma will ignore problems with the given IDs
for the current package, starting from the point where
the pragma is issued.

The parser is very strict and does not accept any whitespace in
"#catkin_lint:". Malformed pragmas are treated like comments and
silently ignored.

Fixes #7