Handle changed ntpath.isabs behaviour in Python 3.13 (#110)
ntpath.isabs no longer considers paths that start with "/" to
be absolute. This is correct behaviour on Windows but causes us
problems, because our `PathConstants` use paths that start with
/ and expect these to always be considered absolute. To deal
with this, let's add an extra slash to the `PathConstants`;
paths starting `//` are considered absolute by both ntpath and
posixpath.

Signed-off-by: Adam Williamson <awilliam@redhat.com>